First commid: readme and minimum-viable-model
Have added a README.md with some initial descriptions, although this
is expected to change a lot.

Have added a file called `bp-v1-mvm.json` which is the minimum  viable
model for the botech protocol (bp) version 1. The implication is that
this should pass all tests written for botech utilities, and that
nothing simpler should.