Added `bp-v1-full.json`, changed "author" to "authors" as a list
I added `bp-v1-full.json` which is expected to be the full
implementation of the botech protocol v1.0.

Furthermore, in doing so, I changed the `author` key to `authors`, and
changed the filetype to a list, which is expected to be a list of
strings, representing the authors of each Botech file. Previously it
was only a string, which meant only one author could be attributed appropriately.